Reconnect after server disconnect or app suspends
Previously when the app lost connection (because either the backend
closed the connection, or because a front react-native app suspended
and then resumed) we were trying to join a channel twice.

Turns out the Phoenix.js already handles re-connection, so if we
have joined the channel previously, we just keep sending the messages
as Phoenix.js will enqueue them and send upon reconnection.

Refactored code to be promise-pipeable.